+++ app/src/main/java/com/example/task23/data/ServiceProvider.kt
@@ -1,0 +1,30 @@
+package com.example.task23.data
+
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+private const val BASEURL = "https://api.openweathermap.org/"
+
+object ServiceProvider {
+
+    fun getWeatherService(): WeatherService {
+        val retrofit = Retrofit
+            .Builder()
+            .client(okHttpClient())
+            .baseUrl(BASEURL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+        retrofit.create(WeatherService::class.java)
+        return retrofit.create(WeatherService::class.java)
+    }
+
+    private fun okHttpClient(): OkHttpClient {
+        val logging = HttpLoggingInterceptor()
+        logging.setLevel(HttpLoggingInterceptor.Level.BODY)
+        return OkHttpClient.Builder()
+            .addInterceptor(logging)
+            .build()
+    }
+}
